# cmd/swaggerui/dist/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This documentation describes a demo web server based on Gin APIs and code will be found under https://github.com/VeinFu/go-gin-ws",
    "title": "Demo Web Server based on Gin",
    "contact": {
      "name": "vien.fu",
      "email": "vien.fu@ucloud.cn"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/test": {
      "get": {
        "description": "Just a test api",
        "tags": [
          "test"
        ],
        "operationId": "getTestAPI",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "APIFeature",
            "name": "api_feature",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TestAPIResponse"
          },
          "default": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    },
    "/api/v1/users": {
      "post": {
        "description": "Create a new user",
        "tags": [
          "users"
        ],
        "operationId": "createUser",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/userCreateReqParams"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UserResponse"
          },
          "default": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    },
    "/api/v1/users/{name}": {
      "get": {
        "description": "just a desc",
        "tags": [
          "users"
        ],
        "summary": "Get specified user",
        "operationId": "getUser",
        "parameters": [
          {
            "type": "string",
            "description": "User Name",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UserResponse"
          },
          "400": {
            "$ref": "#/responses/ResponseError"
          }
        }
      }
    }
  },
  "definitions": {
    "NewTestAPIResponse": {
      "type": "object",
      "properties": {
        "desc": {
          "type": "string",
          "x-go-name": "Description"
        }
      },
      "x-go-package": "go-gin-ws/models"
    },
    "NewUserResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/UserResponse"
        }
      },
      "x-go-package": "go-gin-ws/models"
    },
    "ResponseError": {
      "description": "ResponseError represent the response error struct",
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "go-gin-ws/models"
    },
    "UserResponse": {
      "type": "object",
      "properties": {
        "address": {
          "description": "user current living address",
          "type": "string",
          "x-go-name": "Address"
        },
        "mobile": {
          "description": "user mobile phone number",
          "type": "string",
          "x-go-name": "Mobile"
        },
        "name": {
          "description": "user name",
          "type": "string",
          "x-go-name": "Name"
        },
        "sex": {
          "description": "user sex",
          "type": "string",
          "x-go-name": "Sex"
        }
      },
      "x-go-package": "go-gin-ws/models"
    },
    "userCreateReqParams": {
      "type": "object",
      "properties": {
        "address": {
          "description": "User Address",
          "type": "string",
          "x-go-name": "Address"
        },
        "mobile": {
          "description": "User Mobile Phone Number",
          "type": "string",
          "x-go-name": "Mobile"
        },
        "name": {
          "description": "User Name",
          "type": "string",
          "x-go-name": "Name"
        },
        "sex": {
          "description": "User Sex",
          "type": "string",
          "x-go-name": "Sex"
        }
      },
      "x-go-package": "go-gin-ws/cmd"
    }
  },
  "responses": {
    "ResponseError": {
      "description": "ResponseError is an error that is used when the request fail",
      "schema": {
        "$ref": "#/definitions/ResponseError"
      }
    },
    "TestAPIResponse": {
      "description": "TestAPIResponse retrieve test result",
      "schema": {
        "$ref": "#/definitions/NewTestAPIResponse"
      }
    },
    "UserResponse": {
      "description": "UserResponse retrieve specified user",
      "schema": {
        "$ref": "#/definitions/NewUserResponse"
      }
    }
  }
}